create usePhaseForm hook for phase form, moved all forms to a /forms folder, moved useWindowSize to its own file
