+++ src/components/Project/ProjectNotes/NewNoteForm.tsx
@@ -1,5 +1,5 @@
 import { useAppDispatch } from '@/hooks/common';
-import useProjectNoteForm from '@/hooks/useNoteForm';
+import useProjectNoteForm from '@/forms/useNoteForm';
 import { IProjectNoteForm } from '@/interfaces/formInterfaces';
 import { INoteRequest } from '@/interfaces/noteInterfaces';
 import { postNoteThunk } from '@/reducers/noteSlice';
