+++ src/hooks/common.ts
@@ -1,19 +1,6 @@
-import { useLayoutEffect, useState } from 'react';
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 import type { RootState, AppDispatch } from '../store';
 
 // Use throughout the app instead of plain `useDispatch` and `useSelector` for typed hooks
 export const useAppDispatch: () => AppDispatch = useDispatch;
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
-
-// Listen to window resizes, useful for doing something with breakpoints outside of css files
-export const useWindowSize = () => {
-  const [size, setSize] = useState([0, 0]);
-  useLayoutEffect(() => {
-    const updateSize = () => setSize([window.innerWidth, window.innerHeight]);
-    window.addEventListener('resize', updateSize);
-    updateSize();
-    return () => window.removeEventListener('resize', updateSize);
-  }, []);
-  return size;
-};
