+++ src/hooks/useWindowSize.ts
@@ -1,0 +1,15 @@
+import { useLayoutEffect, useState } from 'react';
+
+// Listen to window resizes, useful for doing something with breakpoints outside of css files
+const useWindowSize = () => {
+  const [size, setSize] = useState([0, 0]);
+  useLayoutEffect(() => {
+    const updateSize = () => setSize([window.innerWidth, window.innerHeight]);
+    window.addEventListener('resize', updateSize);
+    updateSize();
+    return () => window.removeEventListener('resize', updateSize);
+  }, []);
+  return size;
+};
+
+export default useWindowSize;
